Rename component and create test file
