+++ frontend/src/components/Observables/ObservableActions/UpdateDetectionExpiration.vue
